Extract delivery context logic from fetch.py and utils.py into delivery_context.py
