+++ saleor/graphql/checkout/mutations/checkout_shipping_method_update.py
@@ -1,12 +1,15 @@
 import graphene
 from django.core.exceptions import ValidationError
 
+from ....checkout.delivery_context import (
+    assign_delivery_method_to_checkout,
+    get_or_fetch_checkout_deliveries,
+)
 from ....checkout.error_codes import CheckoutErrorCode
 from ....checkout.fetch import (
     CheckoutInfo,
     fetch_checkout_info,
     fetch_checkout_lines,
-    get_or_fetch_checkout_deliveries,
 )
 from ....checkout.models import CheckoutDelivery
 from ....checkout.utils import (
@@ -26,7 +29,6 @@ from ...plugins.dataloaders import get_plugin_manager_promise
 from ..types import Checkout
 from .utils import (
     ERROR_DOES_NOT_SHIP,
-    assign_delivery_method_to_checkout,
     get_checkout,
 )
